requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
python-dotenv
google-generativeai
flask-mail
